requirements.txt
google-api-python-client
google-auth
tabulate
colorama
